% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-ACL.R
\name{ACL}
\alias{ACL}
\title{Create an Access Control List for an Parse object}
\usage{
ACL(public_read = TRUE, public_write = TRUE, owner_write = TRUE)
}
\arguments{
\item{public_read}{whether anyone can read this object}

\item{public_write}{whether anyone can modify this object}

\item{owner_write}{whether the currently logged-in user can modify
this object}
}
\value{
An "ACL" object, which is a list containing permissions
for the current user and/or the public ("*"). This object
can be passed as an \code{ACL =} argument to \code{parse_object}.
}
\description{
An Access Control List determines who can read and write a particular
object. In particular, it chooses whether there is public read or
write access. The user could also choose to deny themselves
write access.
}
\details{
The user must be logged in to create an ACL object.
}
\examples{
\dontrun{

# save some scores from a game
parse_login("myaccount", "mypassword")

# public:
game_score <- parse_object("GameScore", score = 42)

# closed to public writing:
nowrite_acl <- ACL(public_write = FALSE)
game_score2 <- parse_object("GameScore", score = 43, ACL = nowrite_acl)

# closed to public reading or writing:
noread_acl <- ACL(public_read = FALSE, public_write = FALSE)
game_score3 <- parse_object("GameScore", score = 44, ACL = noread_acl)

}
}

