% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-batch.R
\name{as.parse_batch.data.frame}
\alias{as.parse_batch.data.frame}
\title{convert a data.frame to a batch of Parse objects}
\usage{
\method{as.parse_batch}{data.frame}(x, class_name, ...)
}
\arguments{
\item{x}{data.frame to be converted}

\item{class_name}{class name of parse objects to be created}

\item{...}{extra arguments, not used}
}
\description{
convert a data.frame to a batch of Parse objects
}

