% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-batch.R
\name{as.parse_batch}
\alias{as.parse_batch}
\title{convert an object to a batch of Parse objects}
\usage{
as.parse_batch(x, ...)
}
\arguments{
\item{x}{object to be converted}

\item{...}{extra arguments}
}
\description{
convert an object to a batch of Parse objects
}

