% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-object.R
\name{as.parse_object.list}
\alias{as.parse_object.list}
\title{coerce a list into a parse_object}
\usage{
\method{as.parse_object}{list}(x, class_name, ...)
}
\arguments{
\item{x}{list to be coerced}

\item{class_name}{class name to convert to}

\item{...}{extra arguments, not used}
}
\description{
coerce a list into a parse_object
}

