% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-object.R
\name{as.parse_object}
\alias{as.parse_object}
\title{coerce an object into a parse_object}
\usage{
as.parse_object(x, ...)
}
\arguments{
\item{x}{object to be coerced}

\item{...}{extra arguments}
}
\description{
coerce an object into a parse_object
}

