% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-pointer.R
\name{as.parse_pointer.default}
\alias{as.parse_pointer.default}
\title{convert an ordinary object to a Parse pointer}
\usage{
\method{as.parse_pointer}{default}(x, ...)
}
\arguments{
\item{x}{ordinary object}

\item{...}{extra arguments, not used}
}
\description{
This does nothing to the object
}

