% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-file.R
\name{as.parse_pointer.parse_file}
\alias{as.parse_pointer.parse_file}
\title{convert a Parse file to a pointer}
\usage{
\method{as.parse_pointer}{parse_file}(x, ...)
}
\arguments{
\item{x}{parse_file to be converted}

\item{...}{extra arguments (not used)}
}
\description{
convert a Parse file to a pointer
}

