% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-pointer.R
\name{as.parse_pointer.parse_object}
\alias{as.parse_pointer.parse_object}
\title{convert a Parse object to a Parse pointer}
\usage{
\method{as.parse_pointer}{parse_object}(x, ...)
}
\arguments{
\item{x}{parse_object to convert}

\item{...}{extra arguments, not used}
}
\description{
Note that the Parse object must already be saved
}

