% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-pointer.R
\name{as.parse_pointer}
\alias{as.parse_pointer}
\title{convert an object to a Parse pointer}
\usage{
as.parse_pointer(x, ...)
}
\arguments{
\item{x}{object to convert}

\item{...}{extra arguments}
}
\description{
Works on parse_object and parse_file
}

