% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{assert_object_id}
\alias{assert_object_id}
\title{assert that a class has an object ID}
\usage{
assert_object_id(x)
}
\arguments{
\item{x}{A Parse object}
}
\description{
assert that a class has an object ID
}

