% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{batch_body}
\alias{batch_body}
\title{turn multiple columns into a batch set of requests}
\usage{
batch_body(method, path, body = NULL)
}
\arguments{
\item{method}{HTTP method such as GET, POST, or DELETE}

\item{path}{path within Parse API, or vector of paths}

\item{body}{data frame of body for each request in batch}
}
\description{
turn multiple columns into a batch set of requests
}

