% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-save.R
\name{convert.parse_object}
\alias{convert.parse_object}
\title{prepare a Parse object to be saved or queried, by converting items to pointers}
\usage{
\method{convert}{parse_object}(x, ...)
}
\arguments{
\item{x}{parse_object to be converted}

\item{...}{extra arguments}
}
\description{
prepare a Parse object to be saved or queried, by converting items to pointers
}

