% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-save.R
\name{convert}
\alias{convert}
\title{prepare an object to be saved or queried to Parse, by converting items to pointers}
\usage{
convert(x, ...)
}
\arguments{
\item{x}{object to be converted}

\item{...}{extra arguments}
}
\description{
prepare an object to be saved or queried to Parse, by converting items to pointers
}

