% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-httr.R
\name{parse_api_DELETE}
\alias{parse_api_DELETE}
\title{Perform a DELETE request to parse}
\usage{
parse_api_DELETE(path, ...)
}
\arguments{
\item{path}{path to send DELETE request to}

\item{...}{extra arguments passed on to \code{httr::DELETE}}
}
\description{
Perform a DELETE request to parse
}

