% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-httr.R
\name{parse_api_GET}
\alias{parse_api_GET}
\title{Perform a GET request to parse}
\usage{
parse_api_GET(path, ...)
}
\arguments{
\item{path}{path to send GET request to}

\item{...}{extra arguments passed on to \code{httr::GET}}
}
\description{
Perform a GET request to parse
}

