% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-httr.R
\name{parse_api_POST}
\alias{parse_api_POST}
\title{Perform a POST request to parse}
\usage{
parse_api_POST(path, body = NULL, to_json = TRUE, ...)
}
\arguments{
\item{path}{path to send POST request to}

\item{body}{body of POST request}

\item{to_json}{whether to convert body to JSON}

\item{...}{extra arguments passed on to \code{httr::POST}}
}
\description{
Perform a POST request to parse
}

