% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-httr.R
\name{parse_api_PUT}
\alias{parse_api_PUT}
\title{Perform a PUT request to parse}
\usage{
parse_api_PUT(path, body, to_json = TRUE, ...)
}
\arguments{
\item{path}{path to send PUT request to}

\item{body}{body of PUT request}

\item{to_json}{whether to convert body to JSON}

\item{...}{extra arguments passed on to \code{httr::PUT}}
}
\description{
Perform a PUT request to parse
}

