% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-object.R
\name{parse_class}
\alias{parse_class}
\title{get the class of a Parse object}
\usage{
parse_class(x)
}
\arguments{
\item{x}{parse_object or parse_batch}
}
\description{
get the class of a Parse object
}

