% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-delete.R
\name{parse_delete.parse_batch}
\alias{parse_delete.parse_batch}
\title{delete a batch of Parse objects}
\usage{
\method{parse_delete}{parse_batch}(x, ...)
}
\arguments{
\item{x}{parse_batch to be deleted}

\item{...}{extra arguments to pass on to parse_api_POST}
}
\description{
delete a batch of Parse objects
}

