% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-delete.R
\name{parse_delete.parse_object}
\alias{parse_delete.parse_object}
\title{delete a parse object}
\usage{
\method{parse_delete}{parse_object}(x, ...)
}
\arguments{
\item{x}{parse_object to be deleted}

\item{...}{extra arguments to be passed on to parse_api_DELETE}
}
\description{
delete a parse object
}

