% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-delete.R
\name{parse_delete}
\alias{parse_delete}
\title{delete a parse object or batch of objects}
\usage{
parse_delete(x, ...)
}
\arguments{
\item{x}{Parse object or batch to be deleted}

\item{...}{extra arguments to be passed on to parse_api_DELETE}
}
\description{
delete a parse object or batch of objects
}

