% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testing.R
\name{parse_delete_all}
\alias{parse_delete_all}
\title{functions for setting up a Parse testing environment
Delete all objects of a class}
\usage{
parse_delete_all(class_name)
}
\arguments{
\item{class_name}{one or more classes to delete all objects from}
}
\description{
This is vectorized to take multiple classes. Be sure not to
call this by accident, and note that you need permission to
delete all of these objects.
}
\examples{
\dontrun{
parse_delete_all(c("Car", "GameScore"))
}
}

