% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-file.R
\name{parse_file}
\alias{parse_file}
\title{Upload a file to Parse}
\usage{
parse_file(name, path, type = NULL)
}
\arguments{
\item{name}{name to save the file to}

\item{path}{path to file to upload}

\item{type}{type of file (e.g. "application/zip")}
}
\value{
a parse_file object, containing a name and URL
}
\description{
Upload a file to Parse
}

