% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-cloud.R
\name{parse_function}
\alias{parse_function}
\title{Create an interface to a Parse CloudCode function}
\usage{
parse_function(name)
}
\arguments{
\item{name}{Name of the function in CloudCode}
}
\description{
Create an interface to a Parse CloudCode function. This would have to have
been added to your "cloud" directory in your Parse application. For more,
see here: https://parse.com/docs/cloud_code_guide
}
\examples{
# suppose a "hello world" function has been set up in the cloud

hello_function <- parse_function("hello")
hello_function()
}

