% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-user.R
\name{parse_login}
\alias{parse_login}
\title{Log a user into Parse}
\usage{
parse_login(username, password)
}
\arguments{
\item{username}{username}

\item{password}{password}
}
\value{
A _User parse_object
}
\description{
This logs a user into Parse based on a username and pasword combination.
}
\details{
A user's login token and other info is stored in the
\code{parse_user} option.
}

