% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-object.R
\name{parse_object}
\alias{parse_object}
\title{create and save a parse object}
\usage{
parse_object(class_name, ..., save_after = TRUE)
}
\arguments{
\item{class_name}{name of the class}

\item{...}{fields contained in the object}

\item{save_after}{whether to s}
}
\description{
Construct a parse_object with the given fields. By default, saves the object to
Parse afterwards.
}

