% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-user.R
\name{parse_password_reset}
\alias{parse_password_reset}
\title{reset a user's password (if they have an email address registered)}
\usage{
parse_password_reset(email)
}
\arguments{
\item{email}{User's e-mail address}
}
\description{
reset a user's password (if they have an email address registered)
}

