% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-query.R
\name{parse_query}
\alias{parse_query}
\title{retrieve 1000 items from an object based on a query}
\usage{
parse_query(class_name, object_id, limit = 1000, skip = 0, ...)
}
\arguments{
\item{class_name}{name of Parse class}

\item{object_id}{if given, used to retrieve an object}

\item{skip}{if given, specifies how many items to skip before first returned item}

\item{...}{extra arguments, not used}
}
\value{
a Parse object
}
\description{
retrieve 1000 items from an object based on a query
}

