% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-query.R
\name{parse_queryAll}
\alias{parse_queryAll}
\title{get all records and concatenate the output into a single data frame}
\usage{
parse_queryAll(objName, ...)
}
\arguments{
\item{...}{extra arguments, not used}

\item{class_name}{name of Parse class}

\item{object_id}{if given, used to retrieve an object}
}
\value{
a Parse object
}
\description{
get all records and concatenate the output into a single data frame
}

