% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testing.R
\name{parse_restore_state}
\alias{parse_restore_state}
\title{Restore a testing environment back to its original state}
\usage{
parse_restore_state(to_remove = NULL)
}
\arguments{
\item{to_remove}{one or more classes to delete all objects of in test
environment}
}
\description{
Restore a testing environment created with \code{\link{parse_setup_test}}
back to its original state, based on the backup environment variables
created therein. In the process it can clear one or more classes.
}

