% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-save.R
\name{parse_save.data.frame}
\alias{parse_save.data.frame}
\title{create a batch of Parse objects from a data frame}
\usage{
\method{parse_save}{data.frame}(x, class_name, ...)
}
\arguments{
\item{x}{data.frame to be saved}

\item{class_name}{class to save the objects to}

\item{...}{extra arguments to pass on to parse_api_POST}
}
\description{
create a batch of Parse objects from a data frame
}

