% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-save.R
\name{parse_save.list}
\alias{parse_save.list}
\title{create a Parse object from a list}
\usage{
\method{parse_save}{list}(x, class_name, ...)
}
\arguments{
\item{x}{list to be saved}

\item{class_name}{class to save the object to}

\item{...}{extra arguments to pass on to parse_api_POST}
}
\description{
create a Parse object from a list
}

