% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-save.R
\name{parse_save.parse_batch}
\alias{parse_save.parse_batch}
\title{create a batch of Parse objects}
\usage{
\method{parse_save}{parse_batch}(x, ...)
}
\arguments{
\item{x}{parse_batch to be saved}

\item{...}{extra arguments to pass on to parse_api_POST}
}
\description{
create a batch of Parse objects
}

