% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-save.R
\name{parse_save.parse_object}
\alias{parse_save.parse_object}
\title{save a parse object}
\usage{
\method{parse_save}{parse_object}(x, ...)
}
\arguments{
\item{x}{parse_object to be saved}

\item{...}{extra arguments to be passed on to parse_api_POST}
}
\description{
save a parse object
}

