% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-save.R
\name{parse_save}
\alias{parse_save}
\title{save a parse object or batch}
\usage{
parse_save(x, ...)
}
\arguments{
\item{x}{object to be saved}

\item{...}{extra arguments}
}
\description{
save a parse object or batch
}

