% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-user.R
\name{parse_signup}
\alias{parse_signup}
\title{create a new user}
\usage{
parse_signup(username, password, ...)
}
\arguments{
\item{username}{desired username}

\item{password}{desired password}

\item{...}{extra arguments to set, such as email}
}
\value{
A _User parse_object
}
\description{
Create and sign in a user
}
\details{
A user's login token and other info is stored in the
\code{parse_user} option.
}

