% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-query.R
\name{parse_query}
\alias{parse_query}
\title{retrieve an object based on a query}
\usage{
parse_query(class_name, object_id, limit = 1000, ...)
}
\arguments{
\item{class_name}{name of Parse class}

\item{object_id}{if given, used to retrieve an object}

\item{...}{extra arguments, not used}
}
\value{
a Parse object
}
\description{
retrieve an object based on a query
}

