% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse-httr.R
\name{process_Parse}
\alias{process_Parse}
\title{process a request object from Parse}
\usage{
process_Parse(req)
}
\arguments{
\item{req}{request object returned from an httr method}
}
\description{
process a request object from Parse
}

